Add extra CDP logging to capture missing request data.
As described in the comments, Puppeteer seems to be missing some headers in requests and responses if you look at the default data it provides. This appears to be because it only listens to the default Chrome DevTools Protocol default request and response events and not the extra ones.

Added code here to capture the regular and extra events data for requests and responses. This is linked with the browser data by the page _requestId which is the same ID used by the CDP data.